requirements.txt
paramiko==3.3.1
pyyaml